requirement.txt
yt-dlp
pillow
requests
brew reinstall tcl-tk
brew install thinker